+++ modin/core/dataframe/algebra/default2pandas/dataframe.py
@@ -13,8 +13,6 @@
 
 """Module houses default DataFrame functions builder class."""
 
-# FIXME: This whole module is duplicating the logic of `default.py` and should be removed.
-
 from .default import DefaultMethod
 from modin.utils import _inherit_docstrings
 
@@ -23,27 +21,4 @@ import pandas
 
 @_inherit_docstrings(DefaultMethod)
 class DataFrameDefault(DefaultMethod):
-    @classmethod
-    def register(cls, func, obj_type=None, **kwargs):
-        """
-        Build function that do fallback to default pandas implementation for passed `func`.
-
-        Parameters
-        ----------
-        func : callable or str,
-            Function to apply to the casted to pandas frame.
-        obj_type : object, optional
-            If `func` is a string with a function name then `obj_type` provides an
-            object to search function in. If not specified `pandas.DataFrame` will be used.
-        **kwargs : kwargs
-            Additional parameters that will be used for building.
-
-        Returns
-        -------
-        callable
-            Function that takes query compiler, does fallback to pandas and applies `func`
-            to the casted to pandas frame.
-        """
-        if obj_type is None:
-            obj_type = pandas.DataFrame
-        return super().register(func, obj_type=obj_type, **kwargs)
+    DEFAULT_OBJECT_TYPE = pandas.DataFrame
